requirementes.txt
aiohttp==3.7.2
altgraph==0.17
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==3.2.1
cssselect==1.1.0
Cython==0.29.21
Django==3.2
django-shortuuidfield==0.1.3
et-xmlfile==1.1.0
future==0.18.2
gevent==20.9.0
greenlet==0.4.17
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
jmespath==0.10.0
MarkupSafe==1.1.1
MouseInfo==0.1.3
multidict==5.0.2
mysqlclient==2.0.3
openpyxl==3.0.7
parsel==1.6.0
pefile==2019.4.18
Pillow==8.0.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pygame==2.0.1
PyGetWindow==0.0.9
PyHamcrest==2.0.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python-memcached==1.59
PyTweening==1.0.3
pytz==2021.1
pywifi==1.1.12
pywin32==300
pywin32-ctypes==0.2.0
queuelib==1.5.0
requests==2.24.0
service-identity==18.1.0
shortuuid==1.0.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
Twisted @ file:///C:/Users/Demacia/Twisted-20.3.0-cp39-cp39-win_amd64.whl
typing-extensions==3.7.4.3
urllib3==1.25.11
w3lib==1.22.0
Werkzeug==1.0.1
yarl==1.6.3
youtube-dl==2021.4.26
zope.event==4.5.0
zope.interface==5.2.0
